requirements.txt
numpy
astropy
matplotlib
Pillow
sigfig
image_registration
pytest
photutils
scikit-image
ccdproc
astroalign